requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.15.1
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
djangorestframework==3.14.0
fonttools==4.42.0
importlib-resources==6.0.1
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.24.4
packaging==23.1
Pillow==10.0.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.7.1
xattr==0.10.1
zipp==3.16.2
